requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
decorator==5.1.1
distro==1.9.0
executing==2.0.1
faicons==0.2.2
fake-useragent==1.5.1
filelock==3.16.1
fsspec==2024.9.0
groq==0.11.0
h11==0.14.0
htmltools==0.5.3
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.2
idna==3.7
ipython==8.18.0
ipywidgets==8.1.3
jedi==0.19.1
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
langchain-core==0.3.10
langchain-groq==0.2.0
langchain-mistralai==0.2.0
langsmith==0.1.133
linkify-it-py==2.0.3
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
numpy==2.0.1
openai==1.51.2
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
platformdirs==4.2.2
plotly==5.23.0
prompt-toolkit==3.0.36
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
questionary==2.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rsconnect_python==1.24.0
semver==2.13.0
shiny==1.0.0
shinywidgets==0.3.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.2
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.11
